spark_processors/requirements.txt
pyspark>=3.4.0
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary==2.9.10
pytest>=7.4.0
python-dateutil>=2.8.2
python-dotenv>=1.1.0
